Disable multicast in nac_bypass_setup.sh
In most tested cases the script ran without any issues. In one use case, the switch took the interface down. We noticed that when plugging in the network cable, an IGMP message to sign up for multicast is sent. This is done with the real MAC of the interface.

The proposed change does disable multicast and hence no IGMP packet is sent.