Setting SWMAC in case the of Connection Setup only
When nac_bypass_setup.sh is called with the parameter -c, the initial setup is not executed. Hence the SWMAC is set to the initial value and the following NAT rules will not work. This leads to communication from the rogue device to the internal network using the MAC of the rogue device and not the client.